% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{available_lfs_pumf_versions}
\alias{available_lfs_pumf_versions}
\title{List available PUMF LFS versions}
\usage{
available_lfs_pumf_versions()
}
\value{
A tibble with versions and urls to available LFS PUMF data
}
\description{
List available PUMF LFS versions
}
