% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{convert_pumf_numeric_columns}
\alias{convert_pumf_numeric_columns}
\title{Convert columns to numeric and convert all missing values to \code{NA}}
\usage{
convert_pumf_numeric_columns(
  pumf_data,
  pumf_base_path = attr(pumf_data, "pumf_base_path"),
  layout_mask = attr(pumf_data, "layout_mask"),
  numeric_columns = NULL,
  set_missing = TRUE,
  na.values = c(".")
)
}
\arguments{
\item{pumf_data}{pumf data file}

\item{pumf_base_path}{optional base path, guessed from attributes on \code{pumf_data}}

\item{layout_mask}{optional layout mask in case there are several layout files,
guessed from attributes on \code{layout_mask}}

\item{numeric_columns}{optional list of columns to conver, guessed from data if none provided}

\item{set_missing}{logical, will set missing values to \code{NA} if set to \code{TRUE} (default)}

\item{na.values}{optional vectors of values to convert to \code{NA}}
}
\value{
data frame with converted values
}
\description{
Convert columns to numeric and convert all missing values to \code{NA}
}
