% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{download_lfs_pumf}
\alias{download_lfs_pumf}
\title{Download PUMF LFS data}
\usage{
download_lfs_pumf(
  version = "2021",
  pumf_cache_path = getOption("canpumf.cache_path"),
  timeout = 3000
)
}
\arguments{
\item{version}{A version of the pumf data of fornat <Year>-<Month>, e.g. "2021-01"}

\item{pumf_cache_path}{Optional path where to store the extracted PUMF data, default is `getOption("canpumf.cache_path")`}

\item{timeout}{Optional parameter to specify connection timeout for download}
}
\value{
pumf_base_dir that can be used in the other package functions
}
\description{
Download PUMF LFS data
}
