% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{download_pumf}
\alias{download_pumf}
\title{Download PUMF data (deprecated)}
\usage{
download_pumf(
  path,
  destination_dir = file.path(tempdir(), "pumf"),
  refresh = FALSE,
  timeout = 3000
)
}
\arguments{
\item{path}{Download URL for the PUMF zip file.}

\item{destination_dir}{Directory to extract into.}

\item{refresh}{Force re-download even if directory exists.}

\item{timeout}{Connection timeout in seconds.}
}
\value{
pumf_base_dir that can be used in the other package functions
}
\description{
`r lifecycle::badge("deprecated")`

Use [pumf_locate_or_download()] internally or [get_pumf()] as the public
entry point.
}
\keyword{internal}
