% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_pumf_miss_labels}
\alias{get_pumf_miss_labels}
\title{Get PUMF missing data information}
\usage{
get_pumf_miss_labels(pumf_base_path, miss_path = NULL)
}
\arguments{
\item{pumf_base_path}{pumf base path}

\item{miss_path}{optional path or mask for the layout file in case there are several,}
}
\value{
tibble with missing data information
}
\description{
Get PUMF missing data information
}
