% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_pumf_val_labels}
\alias{get_pumf_val_labels}
\title{Get PUMF value labels}
\usage{
get_pumf_val_labels(pumf_base_path, val_path = NULL)
}
\arguments{
\item{pumf_base_path}{pumf base path}

\item{val_path}{optional path or mask for the layout file in case there are several,}
}
\value{
tibble with value labels
}
\description{
Get PUMF value labels
}
