% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_pumf_var_labels}
\alias{get_pumf_var_labels}
\title{Get PUMF variable labels}
\usage{
get_pumf_var_labels(pumf_base_path, var_path = NULL)
}
\arguments{
\item{pumf_base_path}{pumf base path}

\item{val_path}{optional path or mask for the layout file in case there are several,}
}
\value{
tibble with variable labels
}
\description{
Get PUMF variable labels
}
