% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{guess_numeric_pumf_columns}
\alias{guess_numeric_pumf_columns}
\title{Guess which columns in pumf data are numeric}
\usage{
guess_numeric_pumf_columns(
  pumf_base_path,
  layout_mask = NULL,
  numeric_pattern = "\\\\d+-\\\\d+|THRU 99"
)
}
\arguments{
\item{pumf_base_path}{pumf base path}

\item{layout_mask}{optional layout mask in case there are several layout files,}

\item{numeric_pattern}{optional pattern to guess numeric columns from their \code{NA} value range}
}
\value{
vector of column names
}
\description{
Guess which columns in pumf data are numeric
}
