% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{list_pumf_collection}
\alias{list_pumf_collection}
\title{List StatCan PUMF collection}
\usage{
list_pumf_collection()
}
\value{
a tibble with a list of the StatCan PUMF collection
}
\description{
While all of these PUMF files are in principle available,
not all are available through this package, mostly because there is no public download link.
}
