% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cards.R
\name{parse_pumf_data_cards}
\alias{parse_pumf_data_cards}
\title{Parse PUMF data}
\usage{
parse_pumf_data_cards(pumf_base_path, layout_mask = NULL)
}
\arguments{
\item{pumf_base_path}{path for raw pumf data}

\item{layout_mask}{layout mask}
}
\value{
tibble with variable labels
}
\description{
Parse PUMF data
}
\keyword{internal}
