% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_csv.R
\name{parse_pumf_data_csv}
\alias{parse_pumf_data_csv}
\title{Parse PUMF data}
\usage{
parse_pumf_data_csv(data_base_path, pumf_path)
}
\arguments{
\item{data_base_path}{path for raw pumf data}

\item{pumf_path}{path for ourput pumf data}
}
\value{
tibble with variable labels
}
\description{
Parse PUMF data
}
