% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_csv.R
\name{parse_pumf_metadata_csv}
\alias{parse_pumf_metadata_csv}
\title{Parse PUMF medatadata}
\usage{
parse_pumf_metadata_csv(pumf_base_path)
}
\arguments{
\item{pumf_base_path}{path for pumf data}
}
\description{
Parse PUMF medatadata
}
