% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_spss.R
\name{parse_pumf_metadata_spss}
\alias{parse_pumf_metadata_spss}
\title{Parse PUMF medatadata}
\usage{
parse_pumf_metadata_spss(pumf_base_path, layout_mask = NULL)
}
\arguments{
\item{pumf_base_path}{path for pumf data}

\item{layout_mask}{optional parameter to identify layour files if there are many}
}
\description{
Parse PUMF medatadata
}
