% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{pumf_registry_keys}
\alias{pumf_registry_keys}
\title{List all registered survey keys}
\usage{
pumf_registry_keys()
}
\value{
character vector of `"series/version"` keys
}
\description{
List all registered survey keys
}
\keyword{internal}
