% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{read_pumf_var_labels}
\alias{read_pumf_var_labels}
\title{Read PUMF variable labels}
\usage{
read_pumf_var_labels(pumf_base_path, var_path = NULL)
}
\arguments{
\item{pumf_base_path}{pumf base path}

\item{var_path}{optional path or mask for the layout file in case there are several,}
}
\value{
tibble with variable labels
}
\description{
Read PUMF variable labels
}
