% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{download_lfs_pumf}
\alias{download_lfs_pumf}
\title{Download PUMF LFS data}
\usage{
download_lfs_pumf(
  version = "2021-01",
  destination_dir = file.path(tempdir(), "pumf"),
  timeout = 3000
)
}
\arguments{
\item{version}{A version of the pumf data of fornat <Year>-<Month>, e.g. "2021-01"}

\item{destination_dir}{Optional path where to store the extracted PUMF data, default is `file.path(tempdir(),"pumf")`}

\item{timeout}{Optional parameter to specify connection timout for download}
}
\value{
pumf_base_dir that can be used in the other package functions
}
\description{
Download PUMF LFS data
}
