% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{download_pumf}
\alias{download_pumf}
\title{Download PUMF data}
\usage{
download_pumf(
  path,
  destination_dir = file.path(tempdir(), "pumf"),
  refresh = FALSE,
  timeout = 3000
)
}
\arguments{
\item{path}{Download path for PUMF SPSS data}

\item{destination_dir}{Optional path where to store the extracted PUMF data, default is `file.path(tempdir(),"pumf")`}

\item{refresh}{Optional parameter to force re-download of PUMF data}

\item{timeout}{Optional parameter to specify connection timeout for download}
}
\value{
pumf_base_dir that can be used in the other package functions
}
\description{
Download PUMF data
}
\keyword{internal}
