% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{download_pumf}
\alias{download_pumf}
\title{Download PUMF data}
\usage{
download_pumf(path, destination_dir = tempdir())
}
\arguments{
\item{path}{download path for PUMF SPSS data}

\item{destination_dir}{optional path where to store the extracted PUMF data}
}
\value{
pumf_base_dir that can be used in the other package functions
}
\description{
Download PUMF data
}
