% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canpumf_duckdb.R
\name{get_pumf_connection}
\alias{get_pumf_connection}
\title{Get select pumf database connections}
\usage{
get_pumf_connection(
  pumf_series,
  pumf_version = NULL,
  layout_mask = NULL,
  file_mask = layout_mask,
  guess_numeric = TRUE,
  pumf_cache_path = getOption("canpumf.cache_path"),
  refresh = FALSE,
  refresh_layout = FALSE,
  timeout = 3000
)
}
\arguments{
\item{pumf_series}{sereis for the pumf data, like LSF, or CHS}

\item{pumf_version}{In case there are several versions of a given series, like for LFS, the version}

\item{layout_mask}{optional layout mask in case there are several files.
identifiers. For LFS this is the month/year.}

\item{file_mask}{optional additional mask to filter down to specific PUMF file if there are several}

\item{guess_numeric}{logical, will guess numeric columns and covert to numeric and set missing values
to \code{NA} if set to \code{TRUE} (default)}

\item{pumf_cache_path}{A path to a permanent cache. If none is fould the data is stored in the temporary
directory for the duration of the session.}

\item{refresh}{optionall re-downlad pumf data, only for series that can be downloaded directly from StatCan}

\item{refresh_layout}{(optional) regenerate the layout and metadata}

\item{timeout}{Optional parameter to specify connection timeout for download}
}
\value{
A database connection
}
\description{
This is a convenience function that downloads and accesses pumf data for
a curated set of pumf datasets.
}
