% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{label_pumf_columns}
\alias{label_pumf_columns}
\title{Rename to human readable column names}
\usage{
label_pumf_columns(
  pumf_data,
  pumf_base_path = attr(pumf_data, "pumf_base_path"),
  layout_mask = attr(pumf_data, "layout_mask")
)
}
\arguments{
\item{pumf_data}{pumf data file}

\item{pumf_base_path}{optional base path, guessed from attributes on \code{pumf_data}}

\item{layout_mask}{optional layout mask in case there are several layout files,
guessed from attributes on \code{layout_mask}}
}
\value{
data frame with renamed columns
}
\description{
Rename to human readable column names
}
