% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{label_pumf_data}
\alias{label_pumf_data}
\title{Add variable labels and rename to human readable column names}
\usage{
label_pumf_data(
  pumf_data,
  pumf_base_path = attr(pumf_data, "pumf_base_path"),
  layout_mask = attr(pumf_data, "layout_mask")
)
}
\arguments{
\item{pumf_data}{pumf data file}

\item{pumf_base_path}{optional base path, guessed from attributes on \code{pumf_data}}

\item{layout_mask}{optional layout mask in case there are several layout files,
guessed from attributes on \code{layout_mask}}
}
\value{
relabeled data frame
}
\description{
Add variable labels and rename to human readable column names
}
