% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf_collection.R
\name{list_canpumf_collection}
\alias{list_canpumf_collection}
\title{List StatCan PUMF collection with canpumf wrappers}
\usage{
list_canpumf_collection()
}
\value{
a tibble with a list of the PUMF files with canpumf convenience wrappers
}
\description{
A list of pumf collections and versions with convenience wrappers in canpumf.
}
