% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumf.R
\name{read_pumf_data}
\alias{read_pumf_data}
\title{Parse PUMF data}
\usage{
read_pumf_data(
  pumf_base_path,
  layout_mask = NULL,
  file_mask = layout_mask,
  guess_numeric = TRUE
)
}
\arguments{
\item{pumf_base_path}{optional base path, guessed from attributes on \code{pumf_data}}

\item{layout_mask}{optional layout mask in case there are several layout files,
guessed from attributes on \code{layout_mask}}

\item{file_mask}{optional additional mask to filter down to specific PUMF file if there are several}

\item{guess_numeric}{logical, will guess numeric columns and covert to numeric and set missing values
to \code{NA} if set to \code{TRUE} (default)}
}
\value{
data frame with one row for each case in the PUMF data
}
\description{
Parse PUMF data
}
