% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{read_pumf_val_labels}
\alias{read_pumf_val_labels}
\title{Get PUMF value labels}
\usage{
read_pumf_val_labels(pumf_base_path, layout_mask = NULL)
}
\arguments{
\item{pumf_base_path}{pumf base path}

\item{layout_mask}{optional path or mask for the layout file in case there are several,}
}
\value{
tibble with value labels
}
\description{
Get PUMF value labels
}
